apps/api/requirements.txt
async-timeout==5.0.1
bcrypt==4.1.2
blinker==1.9.0
boto3==1.34.10
botocore==1.34.162
click==8.1.8
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.1
s3transfer==0.10.4
six==1.17.0
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.23.0
